More predictable fire
Fire will now burn in every direction it's possible to burn
so, similar to the ameoba you can guarantee it will burn
something. Like the ameoba it's chance to burn something
is random so the only unpredictability is how long it
will take to burn.

+++ src/generate-level.js
@@ -30,7 +30,7 @@ export function randomLevel(settings) {
 
   const mapBuffer = new ArrayBuffer(mapArea * 4);
   const map = new Uint32Array(mapBuffer);
-  const mapFlags = new Uint8Array(mapArea);
+  const mapFlags = new Uint16Array(mapArea);
   map.fill(kSymDirt);
 
   for (let i = 0; i < mapWidth; i++) {
